Fixed builder use-after-free
The rewriters were not keeping reference to their associated builders,
so if they were GC'd between their creation and their usage, the program
would abort (at best).

Now it is kept in the uservalue associated with the rewriter userdata. This
forces to create a table as we also need to keep the callback function...
This will be made more efficient in Lua 5.4 as it allows multiple uservalues.